Move token-to-line-token grouping to lexing transformation
Refactors line token grouping from parser to lexing pipeline.

Changes:
- Add TokenStream::Grouped variant with GroupedTokens and GroupType
- Create LineTokenGroupingMapper transformation (from tree_builder logic)
- Update StreamMapper trait with map_grouped() method
- Simplify tree_builder.rs to only handle hierarchy building
- Update linebased parser to accept Vec<LineToken> instead of flat tokens
- Add temporary shim in parser for backward compatibility

Related to #195
