+++ src/lex/lexing/transformations.rs
@@ -4,9 +4,11 @@
 //! that perform specific transformations on TokenStreams.
 
 pub mod blank_lines;
+pub mod line_token_grouping;
 pub mod normalize_whitespace;
 pub mod semantic_indentation;
 
 pub use blank_lines::BlankLinesMapper;
+pub use line_token_grouping::LineTokenGroupingMapper;
 pub use normalize_whitespace::NormalizeWhitespaceMapper;
 pub use semantic_indentation::SemanticIndentationMapper;
